lio_standalone/requirements.txt
numpy>=1.23
scipy>=1.9
numba>=0.57
rosbags>=0.9.20
tqdm>=4.60
pyyaml>=6.0
